repos/tigerinvests-com__sn107-alpha/requirements.txt
bittensor==9.9.0
fastapi==0.110.3
uvicorn[standard]==0.34.0
aiohttp==3.10.11
pydantic==2.11.5
python-dotenv==1.1.0
jsonschema>=4.21,<5
httpx==0.28.1
